To support new kinds of hardisk
With this patch each non-removable hard disk will be recognized.

And partition that is not formated will be listed this time. With
old code partition that is not formated will not be listed accor-
ding to blkid's issue.

Disk size and partition size is shown with MB unit than sectors or
KB.

Partition /system and /data are labeled as oto_sys and oto_data.